+++ screens/AppNavigator/screens/TransactionsScreen/screens/TransactionsScreen.tsx
@@ -7,7 +7,7 @@ import { useWalletAPI } from '../../../../../hooks/wallet/WalletAPI'
 import { useWhaleApiClient } from '../../../../../hooks/api/useWhaleApiClient'
 import { Ionicons } from '@expo/vector-icons'
 import { Text, NumberText } from '../../../../../components'
-import { activitiesToViewModel, VMTransaction } from './statePropcessor'
+import { activitiesToViewModel, VMTransaction } from './stateProcessor'
 import { NavigationProp, useNavigation } from '@react-navigation/native'
 import { TransactionsParamList } from '../TransactionsNavigator'
 import { PrimaryColorStyle } from '../../../../../constants/Theme'
