+++ screens/AppNavigator/screens/TransactionsScreen/screens/stateProcessor.ts
